+++ trunk/src/app/srs_app_hybrid.hpp
@@ -15,6 +15,7 @@
 
 class SrsServer;
 class SrsServerAdapter;
+class SrsWaitGroup;
 
 // The hibrid server interfaces, we could register many servers.
 class ISrsHybridServer
@@ -26,7 +27,7 @@ public:
     // Only ST initialized before each server, we could fork processes as such.
     virtual srs_error_t initialize() = 0;
     // Run each server, should never block except the SRS master server.
-    virtual srs_error_t run() = 0;
+    virtual srs_error_t run(SrsWaitGroup* wg) = 0;
     // Stop each server, should do cleanup, for example, kill processes forked by server.
     virtual void stop() = 0;
 };